@@ -5,20 +5,21 @@ repository: https://github.com/domain-centric/documentation_builder
 documentation: https://github.com/domain-centric/documentation_builder/wiki
 
 environment:
-  sdk: ">=2.12.0 <4.0.0"
+  sdk: '>=3.1.0 <4.0.0'
 
 dependencies:
   build: ^2.3.1
   build_config: ^1.1.0
   build_runner: ^2.2.1
   test: ^1.21.6
-  collection: ^1.16.0
+  collection: ^1.19.1
   fluent_regex: ^1.0.3
   http: ^1.1.0
-  analyzer: ^6.2.0
+  analyzer: ^7.4.4
   logging: ^1.0.2
-  template_engine: ^1.0.0
+  template_engine: ^1.5.1
   shouldly: ^0.5.0+1
-
+  petitparser: ^6.1.0
+  
 dev_dependencies:
-  lints: ^2.0.1
+  lints: ^5.1.1
