@@ -1,597 +1,598 @@
-import 'dart:async';
-import 'dart:io';
-
-import 'package:analyzer/dart/element/element.dart' as analyzer;
-import 'package:build/build.dart';
-import 'package:fluent_regex/fluent_regex.dart';
-
-import '../builder/documentation_builder.dart';
-import '../builder/documentation_model_builder.dart';
-import '../generic/documentation_model.dart';
-import '../generic/element.dart';
-import '../generic/paths.dart';
-import '../project/local_project.dart';
-import 'attribute_parser.dart';
-import 'link_parser.dart';
-import 'parser.dart';
-
-/// The [TagParser] searches for [TextNode]'s that contain texts that represent a [Tag]
-/// It then replaces these [TextNode]'s into a [Tag] and additional [TextNode]'s for the remaining text.
-class TagParser extends Parser {
-  TagParser()
-      : super([
-          ImportFileTagRule(),
-          ImportCodeTagRule(),
-          ImportDartCodeTagRule(),
-          ImportDartDocTagRule(),
-          TitleRule(),
-          TableOfContentsTagRule(),
-          MitLicenseTagRule(),
-        ]);
-}
-
-abstract class TagRule extends TextParserRule {
-  final List<AttributeRule> attributeRules;
-
-  TagRule(String name, this.attributeRules) : super(createExpression(name));
-
-  static FluentRegex createExpression(String name) => FluentRegex()
-      .literal('{')
-      .whiteSpace(Quantity.zeroOrMoreTimes())
-      .literal(name)
-      .group(FluentRegex().anyCharacter(Quantity.zeroOrMoreTimes().reluctant),
-          type: GroupType.captureNamed(GroupName.attributes))
-      .literal('}')
-      .ignoreCase();
-
-  @override
-  Future<Node> createReplacementNode(
-      ParentNode parent, RegExpMatch match) async {
-    try {
-      String attributesText = match.namedGroup(GroupName.attributes) ?? '';
-      var tagAttributeParser = AttributeParser(attributeRules);
-      Map<String, dynamic> attributes =
-          await tagAttributeParser.parseToNameAndValues(attributesText);
-      var tagNode = createTagNode(parent, attributes);
-      var newChildren = await tagNode.createChildren();
-      tagNode.children.addAll(newChildren);
-      return tagNode;
-    } on ParserWarning catch (warning) {
-      // Wrap warning with tag information, so it can be found easily
-      throw ParserWarning("$warning in tag: '${match.result}'.");
-    }
-  }
-
-  Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes);
-}
-
-/// [Tag]s are specific texts in [Template]s that are replaced by the
-///  [DocumentationBuilder] with other information
-///  (e.g. by an imported Dart Documentation Comment) before the
-///  [GeneratedFile] is written.
-///
-/// [Tag]s:
-/// - are surrounded by curly brackets: {}
-/// - start with a name: e.g.  {ImportFile&rcub;
-/// - may have [Attribute]s after the name:
-///   e.g. {ImportFile path='OtherTemplateFile.mdt' title='## Other Template File'&rcub;
-abstract class Tag extends ParentNode {
-  final Map<String, dynamic> attributes;
-  late final Anchor anchor;
-
-  Tag(ParentNode? parent, this.attributes) : super(parent);
-
-  Future<List<Node>> createChildren();
-}
-
-/// - **{ImportFile file:'OtherTemplateFile.mdt' title='## Other Template File'&rcub;**
-/// - Imports another text file or markdown file.
-/// - Attributes:
-///   - path= (required) A [ProjectFilePath] to a file that needs to be imported.
-///     This may be any type of text file (e.g. .mdt file).
-///   - title= (optional) See [TitleAttribute]
-class ImportFileTag extends Tag {
-  ImportFileTag(ParentNode? parent, Map<String, dynamic> attributes)
-      : super(parent, attributes);
-
-  @override
-  Future<List<Node>> createChildren() {
-    ProjectFilePath path = attributes[AttributeName.path];
-    String? title = attributes[AttributeName.title];
-    var titleAndOrAnchor = TitleAndOrAnchor(this, title, path.toString());
-    anchor = titleAndOrAnchor.anchor;
-    var file = path.toFile();
-    var fileText = (TextNode(this, _readFile(file)));
-    return Future.value([titleAndOrAnchor, fileText]);
-  }
-}
-
-/// Recognizes and creates an [ImportFileTag]
-class ImportFileTagRule extends TagRule {
-  ImportFileTagRule()
-      : super('ImportFile', [
-          ProjectFilePathAttributeRule(),
-          TitleAttributeRule(),
-        ]);
-
-  @override
-  Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
-      ImportFileTag(parent, attributes);
-}
-
-/// - **{ImportCodeTag file:'file_to_import.txt' title='## Code example'&rcub;**
-/// - Imports a (none Dart) code file.
-/// - Attributes:
-///   - path= (required) A [ProjectFilePath] a file path that needs to be imported as a (none Dart) code example. See also [ImportDartCodeTag] to import Dart code
-///   - title= (optional) See [TitleAttribute]
-class ImportCodeTag extends Tag {
-  ImportCodeTag(ParentNode? parent, Map<String, dynamic> attributes)
-      : super(parent, attributes);
-
-  @override
-  Future<List<Node>> createChildren() {
-    ProjectFilePath path = attributes[AttributeName.path];
-    String? title = attributes[AttributeName.title];
-    var titleAndOrAnchor = TitleAndOrAnchor(this, title, path.toString());
-    anchor = titleAndOrAnchor.anchor;
-    var codePrefix = TextNode(this, "\n```\n");
-    var file = path.toFile();
-    var fileText = TextNode(this, _readCodeFile(file));
-    var codeSuffix = TextNode(this, "\n```\n");
-    return Future.value([
-      titleAndOrAnchor,
-      codePrefix,
-      fileText,
-      codeSuffix,
-    ]);
-  }
-}
-
-/// Recognizes and creates an [ImportCodeTag]
-class ImportCodeTagRule extends TagRule {
-  ImportCodeTagRule()
-      : super('ImportCode', [
-          ProjectFilePathAttributeRule(),
-          TitleAttributeRule(),
-        ]);
-
-  @override
-  Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
-      ImportCodeTag(parent, attributes);
-}
-
-/// - **{ImportDartCodeTag file:'file_to_import.dart' title='## Dart code example'&rcub;**
-/// - Imports a (none Dart) code file.
-/// - Attributes:
-///   - path= (required) A [DartFilePath] to be imported as a Dart code example. See also [ImportCodeTag] to import none Dart code.
-///   - title= (optional) See [TitleAttribute]
-class ImportDartCodeTag extends Tag {
-  ImportDartCodeTag(ParentNode? parent, Map<String, dynamic> attributes)
-      : super(parent, attributes);
-
-  @override
-  Future<List<Node>> createChildren() async {
-    DartFilePath path = attributes[AttributeName.path];
-    String? title = attributes[AttributeName.title];
-    var titleAndOrAnchor = TitleAndOrAnchor(this, title, path.toString());
-    anchor = titleAndOrAnchor.anchor;
-    var codePrefix = TextNode(this, "\n```dart\n");
-    var code = _readCodeFile(path.toFile());
-    var codeNode = TextNode(this, code);
-    var codeSuffix = TextNode(this, "\n```\n");
-    return [
-      titleAndOrAnchor,
-      codePrefix,
-      codeNode,
-      codeSuffix,
-    ];
-  }
-}
-
-/// Recognizes and creates an [ImportDartCodeTag]
-class ImportDartCodeTagRule extends TagRule {
-  ImportDartCodeTagRule()
-      : super('ImportDartCode', [
-          DartFilePathAttributeRule(),
-          TitleAttributeRule(),
-        ]);
-
-  @override
-  Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
-      ImportDartCodeTag(parent, attributes);
-}
-
-/// - **{ImportDartDoc path='lib\my_lib.dart|MyClass' title='## My Class'&rcub;**
-/// - Imports Dart documentation comments from a library member of a dart file.
-/// - Attributes:
-///   - path= (required) A [DartCodePath] to be imported Dart comments.
-///   - title= (optional) See [TitleAttribute]
-class ImportDartDocTag extends Tag {
-  ImportDartDocTag(ParentNode? parent, Map<String, dynamic> attributes)
-      : super(parent, attributes);
-
-  @override
-  Future<List<Node>> createChildren() async {
-    DartCodePath path = attributes[AttributeName.path];
-    String? title = attributes[AttributeName.title];
-    var titleAndOrAnchor = TitleAndOrAnchor(this, title, path.toString());
-    anchor = titleAndOrAnchor.anchor;
-    var documentation = await _readDocumentationComments(parent!, path);
-    // var documentation = _removeLeadingTripleSlashes(documentationComments);
-    var codeNode = TextNode(this, documentation);
-    return [
-      titleAndOrAnchor,
-      codeNode,
-    ];
-  }
-
-  Future<String> _readDocumentationComments(
-      ParentNode parent, DartCodePath path) async {
-    validate(path);
-
-    analyzer.LibraryElement library =
-        await parseLibrary(parent, path.dartFilePath);
-
-    analyzer.Element foundElement = findAnalyzerElement(library, path);
-
-    var documentationComments = foundElement.documentationComment ?? '';
-    String documentation = _removeLeadingTripleSlashes(documentationComments);
-    validateIfNotEmpty(documentation, path);
-
-    return documentation;
-  }
-
-  static final leadingTripleSlashesExpression = FluentRegex()
-      .startOfLine()
-      .whiteSpace(Quantity.zeroOrMoreTimes())
-      .literal('///')
-      .whiteSpace(Quantity.atMost(1));
-  static final lineBreakExpression = FluentRegex().lineBreak();
-
-  _removeLeadingTripleSlashes(String documentationComments) {
-    var lines = documentationComments.split(lineBreakExpression);
-    String doc = '';
-    lines.forEach((line) {
-      doc += leadingTripleSlashesExpression.removeFirst(line) + '\n';
-    });
-    return doc;
-  }
-
-  void validate(DartCodePath path) {
-    validateIfDartFilePathExists(path);
-    validateIfMemberPathExists(path);
-  }
-
-  void validateIfDartFilePathExists(DartCodePath path) {
-    if (!path.dartFilePath.toFile().existsSync()) {
-      throw ParserWarning(
-          'The Dart file could not be found in path attribute value: $path');
-    }
-  }
-
-  void validateIfMemberPathExists(DartCodePath path) {
-    if (path.dartMemberPath == null) {
-      throw ParserWarning('No DartMemberPath in path attribute value: $path');
-    }
-  }
-
-  analyzer.Element findAnalyzerElement(
-      analyzer.Element element, DartCodePath path) {
-    var visitor = ElementFinder(path.dartMemberPath!);
-    element.visitChildren(visitor);
-
-    analyzer.Element? foundElement = visitor.foundElement;
-    if (foundElement == null) {
-      throw ParserWarning(
-          'Dart member: ${path.dartMemberPath} not found in: ${path.dartFilePath}');
-    }
-    return foundElement;
-  }
-
-  void validateIfMemberFound(
-      analyzer.Element? foundElement, DartCodePath path) {
-    if (foundElement == null) {}
-  }
-
-  void validateIfNotEmpty(String docComments, DartCodePath path) {
-    if (docComments.trim().isEmpty) {
-      throw ParserWarning(
-          'Dart member: ${path.dartMemberPath} has no Dart documentation comments in: ${path.dartFilePath}');
-    }
-  }
-}
-
-Future<analyzer.LibraryElement> parseLibrary(
-    ParentNode parent, ProjectFilePath dartFile) async {
-  try {
-    var documentationModel = parent.findParent<DocumentationModel>();
-    var resolver = documentationModel!.buildStep!.resolver;
-    var assetId = AssetId(LocalProject.name, dartFile.path);
-    var library = await resolver.libraryFor(assetId);
-    return library;
-  } on Exception catch (e) {
-    throw ParserWarning('Could not parse library file: ${dartFile.path}',
-        ParserWarning(e.toString()));
-  }
-}
-
-/// Recognizes and creates an [ImportDartDocTag]
-class ImportDartDocTagRule extends TagRule {
-  ImportDartDocTagRule()
-      : super('ImportDartDoc', [
-          DartCodePathAttributeRule(),
-          TitleAttributeRule(),
-        ]);
-
-  @override
-  Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
-      ImportDartDocTag(parent, attributes);
-}
-
-/// Replaces titles (text starting with # or ## or ### etc) with TitleNodes
-/// This must be done before [TableOfContentsTagRule] so that titles in template files
-/// can also be found by the [TableOfContentsTag]
-class TitleRule extends TextParserRule {
-  TitleRule() : super(createExpression());
-
-  @override
-  Future<Node> createReplacementNode(ParentNode parent, RegExpMatch match) =>
-      Future.value(Title(parent, match.result));
-
-  static FluentRegex createExpression() => FluentRegex()
-      .startOfLine()
-      .whiteSpace(Quantity.zeroOrMoreTimes())
-      .literal('#', Quantity.between(1, 6))
-      .whiteSpace(Quantity.zeroOrMoreTimes())
-      .characterSet(CharacterSet.exclude().addLiterals('#'),
-          Quantity.oneOrMoreTimes().reluctant)
-      .endOfLine();
-}
-
-/// - **{TableOfContents title='## Table of contents example'&rcub;**
-/// - Creates a table of contents.
-/// - Attributes:
-///   - title= (optional) See [TitleAttribute]
-class TableOfContentsTag extends Tag {
-  TableOfContentsTag(ParentNode? parent, Map<String, dynamic> attributes)
-      : super(parent, attributes);
-
-  @override
-  Future<List<Node>> createChildren() {
-    String toc = "";
-    DocumentationModel? model = parent!.findParent<DocumentationModel>();
-    if (model != null) {
-      var templates = model.findOrderedMarkdownTemplates();
-      for (Template template in templates) {
-        if (template is WikiTemplate && !isWikiHomePage(template)) {
-          Link titleLink = Link(
-              parent: parent,
-              title: template.title,
-              uri: template.destinationWebUri!);
-          toc += "\n**$titleLink**\n\n";
-          List<Title> subTitles = template.findChildren<Title>();
-          for (Title subTitle in subTitles) {
-            var title = subTitle.titleWithoutHashes;
-            var uri = subTitle.anchor.uriToAnchor;
-            var level = subTitle.level; //1,2,3
-            if (uri != null && level <= 3) {
-              Link subTitleLink = Link(parent: parent, title: title, uri: uri);
-              toc += "${'  ' * (level - 1)}- $subTitleLink\n";
-            }
-          }
-        }
-      }
-    }
-    String? titleText = attributes[AttributeName.title];
-    var tableOfContents = TextNode(this, toc);
-    if (titleText != null && titleText.trim().isNotEmpty) {
-      var title = Title(this, titleText);
-      anchor = title.anchor;
-      return Future.value([title, tableOfContents]);
-    } else {
-      return Future.value([tableOfContents]);
-    }
-  }
-
-  bool isWikiHomePage(Template template) =>
-      template.sourceFilePath.path.toLowerCase().contains('home.mdt');
-}
-
-/// Recognizes and creates an [TableOfContentsTag]
-class TableOfContentsTagRule extends TagRule {
-  TableOfContentsTagRule()
-      : super('TableOfContents', [
-          TitleAttributeRule(),
-        ]);
-
-  @override
-  Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
-      TableOfContentsTag(parent, attributes);
-}
-
-class TitleAndOrAnchor extends ParentNode {
-  late final Anchor anchor;
-
-  TitleAndOrAnchor(ParentNode? parent, String? title, String path)
-      : super(parent) {
-    if (title != null) {
-      var titleNode = Title(this, title);
-      anchor = titleNode.anchor;
-      children.add(titleNode);
-    } else {
-      anchor = Anchor(this, path);
-      children.add(anchor);
-    }
-  }
-}
-
-/// Recognizes and creates a [MitLicenseTag]
-class MitLicenseTagRule extends TagRule {
-  MitLicenseTagRule()
-      : super('mitLicense',
-            [StringAttributeRule(AttributeName.name, required: true)]);
-
-  @override
-  Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) {
-    return MitLicenseTag(parent, attributes);
-  }
-}
-
-/// - **{MitLicense name='CopyRight holder'&rcub;**
-/// - Writes the MIT License with current year and copyright holder
-/// - Attributes:
-///   - name= (required) Copyright holder name.
-class MitLicenseTag extends Tag {
-  MitLicenseTag(ParentNode? parent, Map<String, dynamic> attributes)
-      : super(parent, attributes);
-
-  String mitLicenseText(String name) => 'MIT License:\n\n'
-      'Copyright (c) ${DateTime.now().year} $name\n\n'
-      'Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:\n\n'
-      'The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.\n\n'
-      'THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.';
-
-  @override
-  Future<List<Node>> createChildren() {
-    String name = attributes[AttributeName.name];
-    return Future.value([TextNode(parent, mitLicenseText(name))]);
-  }
-}
-
-/// Represents a markdown title with a custom Anchor id that can be navigated to using [anchor.uriToAnchor]
-///
-/// Output example:
-/// <a id='paragraph-titile'></a>
-/// ## Paragraph Title
-class Title extends ParentNode {
-  late final Anchor anchor;
-  final String title;
-  final String titleWithoutHashes;
-
-  /// The title can have leading hashtags to indicate the title level, e.g.:
-  /// # = Chapter (level=1)
-  /// ## = Paragraph (level=2)
-  /// ### = Sub paragraph (level=3)
-  /// etc... up to 6 levels.
-  final int level;
-
-  /// The title can have leading hashtags to indicate the title level, e.g.:
-  /// # = Chapter
-  /// ## = Paragraph
-  /// ### = Sub paragraph
-  /// etc... up to 6 levels.
-  Title(ParentNode parent, this.title)
-      : titleWithoutHashes = createTitleWithoutHashes(title),
-        level = createLevel(title),
-        super(parent) {
-    anchor = Anchor(this, title);
-    children.add(anchor);
-    children.add(TitleText(this, title));
-  }
-
-  static final FluentRegex _leadingHashExpression = FluentRegex()
-      .startOfLine()
-      .group(FluentRegex().whiteSpace(Quantity.zeroOrMoreTimes()).literal('#'),
-          quantity: Quantity.atMost(6))
-      .whiteSpace(Quantity.zeroOrMoreTimes());
-
-  static createLevel(String title) {
-    String? leadingHashes = _leadingHashExpression.findFirst(title);
-    if (leadingHashes == null) {
-      return 7;
-    } else {
-      return FluentRegex().literal('#').allMatches(leadingHashes).length;
-    }
-  }
-
-  static createTitleWithoutHashes(String title) {
-    return _leadingHashExpression.removeFirst(title);
-  }
-}
-
-/// [TitleText] is needed because we can not use [TextNode] inside a [Title].
-/// This would create a Stackoverflow by the [TitleRule] that keeps replacing
-/// the [TextNode] inside a [Title]
-class TitleText extends Node {
-  final String title;
-
-  TitleText(ParentNode? parent, this.title) : super(parent);
-
-  /// Titles must be on a single line to be recognized as a title by the markdown engine
-  @override
-  String toString() => '\n$title\n';
-}
-
-/// Represents a HTML anchor point to which you can refer to with a [uriToAnchor]
-/// https://stackoverflow.com/questions/5319754/cross-reference-named-anchor-in-markdown
-class Anchor extends Node {
-  late final String html;
-  late final String name;
-  late final Uri? uriToAnchor;
-  static final firstHyphen = FluentRegex().startOfLine().literal('-');
-  static final multipleHyphen = FluentRegex().literal('-', Quantity.atLeast(2));
-  static final whiteSpace = FluentRegex().whiteSpace();
-  static final otherThanLettersNumbersAndHyphens = FluentRegex().characterSet(
-      CharacterSet.exclude().addLetters().addDigits().addLiterals('-'));
-
-  Anchor(ParentNode? parent, String textToChangeToName) : super(parent) {
-    name = createName(textToChangeToName);
-    html = createHtml(name);
-    uriToAnchor = createUriToAnchor(name);
-  }
-
-  static String createHtml(String name) => "<a id='$name'></a>";
-
-  /// Converts a text to a name that can be referred to in a uri.
-  static String createName(String textToChangeToName) => textToChangeToName
-      .trim()
-      .replaceAll(whiteSpace, '-')
-      .replaceAll(otherThanLettersNumbersAndHyphens, '-')
-      .replaceAll(multipleHyphen, '-')
-      .replaceAll(firstHyphen, '')
-      .toLowerCase();
-
-  Uri? createUriToAnchor(String name) {
-    if (parent == null) return null;
-    Template? markdownTemplate = parent!.findParent<Template>();
-    if (markdownTemplate == null) return null;
-    Uri? uri = markdownTemplate.destinationWebUri;
-    if (uri == null) return null;
-    uri = uri.replace(fragment: name);
-    return uri;
-  }
-
-  @override
-  String toString() => html;
-}
-
-String _readFile(File file) {
-  try {
-    return file.readAsStringSync();
-  } on Exception catch (e) {
-    throw ParserWarning(
-        'Could not read file: ${file.path}.', ParserWarning(e.toString()));
-  }
-}
-
-final _leadingWhiteSpace = FluentRegex()
-    .startOfLine()
-    .whiteSpace(Quantity.oneOrMoreTimes())
-    .multiline(false);
-final _trailingWhiteSpace =
-    FluentRegex().whiteSpace(Quantity.oneOrMoreTimes()).endOfLine();
-
-String _trimWhiteSpace(String text) {
-  text = _leadingWhiteSpace.removeFirst(text);
-  var matches = _trailingWhiteSpace.allMatches(text);
-  if (matches.isNotEmpty) {
-    var lastMatch = matches.last;
-    var endPos = lastMatch.end;
-    if (endPos == text.length) {
-      text = text.substring(0, lastMatch.start);
-    }
-  }
-  return text;
-}
-
-String _readCodeFile(File file) => _trimWhiteSpace(_readFile(file));
+// import 'dart:async';
+// import 'dart:io';
+
+// import 'package:analyzer/dart/element/element.dart' as analyzer;
+// import 'package:build/build.dart';
+// import 'package:documentation_builder/src/builder/new_line.dart';
+// import 'package:fluent_regex/fluent_regex.dart';
+
+// import '../builder/documentation_builder.dart';
+// import '../builder/documentation_model_builder.dart';
+// import '../generic/documentation_model.dart';
+// import '../generic/element.dart';
+// import '../generic/paths.dart';
+// import '../engine/path/local_project.dart';
+// import 'attribute_parser.dart';
+// import 'link_parser.dart';
+// import 'parser.dart';
+
+// /// The [TagParser] searches for [TextNode]'s that contain texts that represent a [Tag]
+// /// It then replaces these [TextNode]'s into a [Tag] and additional [TextNode]'s for the remaining text.
+// class TagParser extends Parser {
+//   TagParser()
+//       : super([
+//           ImportFileTagRule(),
+//           ImportCodeTagRule(),
+//           ImportDartCodeTagRule(),
+//           ImportDartDocTagRule(),
+//           TitleRule(),
+//           TableOfContentsTagRule(),
+//           MitLicenseTagRule(),
+//         ]);
+// }
+
+// abstract class TagRule extends TextParserRule {
+//   final List<AttributeRule> attributeRules;
+
+//   TagRule(String name, this.attributeRules) : super(createExpression(name));
+
+//   static FluentRegex createExpression(String name) => FluentRegex()
+//       .literal('{')
+//       .whiteSpace(Quantity.zeroOrMoreTimes())
+//       .literal(name)
+//       .group(FluentRegex().anyCharacter(Quantity.zeroOrMoreTimes().reluctant),
+//           type: GroupType.captureNamed(GroupName.attributes))
+//       .literal('}')
+//       .ignoreCase();
+
+//   @override
+//   Future<Node> createReplacementNode(
+//       ParentNode parent, RegExpMatch match) async {
+//     try {
+//       String attributesText = match.namedGroup(GroupName.attributes) ?? '';
+//       var tagAttributeParser = AttributeParser(attributeRules);
+//       Map<String, dynamic> attributes =
+//           await tagAttributeParser.parseToNameAndValues(attributesText);
+//       var tagNode = createTagNode(parent, attributes);
+//       var newChildren = await tagNode.createChildren();
+//       tagNode.children.addAll(newChildren);
+//       return tagNode;
+//     } on ParserWarning catch (warning) {
+//       // Wrap warning with tag information, so it can be found easily
+//       throw ParserWarning("$warning in tag: '${match.result}'.");
+//     }
+//   }
+
+//   Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes);
+// }
+
+// /// [Tag]s are specific texts in [Template]s that are replaced by the
+// ///  [DocumentationBuilder] with other information
+// ///  (e.g. by an imported Dart Documentation Comment) before the
+// ///  [GeneratedFile] is written.
+// ///
+// /// [Tag]s:
+// /// - are surrounded by curly brackets: {}
+// /// - start with a name: e.g.  {ImportFile&rcub;
+// /// - may have [Attribute]s after the name:
+// ///   e.g. {ImportFile path='OtherTemplateFile.mdt' title='## Other Template File'&rcub;
+// abstract class Tag extends ParentNode {
+//   final Map<String, dynamic> attributes;
+//   late final Anchor anchor;
+
+//   Tag(ParentNode? parent, this.attributes) : super(parent);
+
+//   Future<List<Node>> createChildren();
+// }
+
+// /// - **{ImportFile file:'OtherTemplateFile.mdt' title='## Other Template File'&rcub;**
+// /// - Imports another text file or markdown file.
+// /// - Attributes:
+// ///   - path= (required) A [ProjectFilePath] to a file that needs to be imported.
+// ///     This may be any type of text file (e.g. .mdt file).
+// ///   - title= (optional) See [TitleAttribute]
+// class ImportFileTag extends Tag {
+//   ImportFileTag(ParentNode? parent, Map<String, dynamic> attributes)
+//       : super(parent, attributes);
+
+//   @override
+//   Future<List<Node>> createChildren() {
+//     ProjectFilePath path = attributes[AttributeName.path];
+//     String? title = attributes[AttributeName.title];
+//     var titleAndOrAnchor = TitleAndOrAnchor(this, title, path.toString());
+//     anchor = titleAndOrAnchor.anchor;
+//     var file = path.toFile();
+//     var fileText = (TextNode(this, _readFile(file)));
+//     return Future.value([titleAndOrAnchor, fileText]);
+//   }
+// }
+
+// /// Recognizes and creates an [ImportFileTag]
+// class ImportFileTagRule extends TagRule {
+//   ImportFileTagRule()
+//       : super('ImportFile', [
+//           ProjectFilePathAttributeRule(),
+//           TitleAttributeRule(),
+//         ]);
+
+//   @override
+//   Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
+//       ImportFileTag(parent, attributes);
+// }
+
+// /// - **{ImportCodeTag file:'file_to_import.txt' title='## Code example'&rcub;**
+// /// - Imports a (none Dart) code file.
+// /// - Attributes:
+// ///   - path= (required) A [ProjectFilePath] a file path that needs to be imported as a (none Dart) code example. See also [ImportDartCodeTag] to import Dart code
+// ///   - title= (optional) See [TitleAttribute]
+// class ImportCodeTag extends Tag {
+//   ImportCodeTag(ParentNode? parent, Map<String, dynamic> attributes)
+//       : super(parent, attributes);
+
+//   @override
+//   Future<List<Node>> createChildren() {
+//     ProjectFilePath path = attributes[AttributeName.path];
+//     String? title = attributes[AttributeName.title];
+//     var titleAndOrAnchor = TitleAndOrAnchor(this, title, path.toString());
+//     anchor = titleAndOrAnchor.anchor;
+//     var codePrefix = TextNode(this, "$newLine```$newLine");
+//     var file = path.toFile();
+//     var fileText = TextNode(this, _readCodeFile(file));
+//     var codeSuffix = TextNode(this, "$newLine```$newLine");
+//     return Future.value([
+//       titleAndOrAnchor,
+//       codePrefix,
+//       fileText,
+//       codeSuffix,
+//     ]);
+//   }
+// }
+
+// /// Recognizes and creates an [ImportCodeTag]
+// class ImportCodeTagRule extends TagRule {
+//   ImportCodeTagRule()
+//       : super('ImportCode', [
+//           ProjectFilePathAttributeRule(),
+//           TitleAttributeRule(),
+//         ]);
+
+//   @override
+//   Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
+//       ImportCodeTag(parent, attributes);
+// }
+
+// /// - **{ImportDartCodeTag file:'file_to_import.dart' title='## Dart code example'&rcub;**
+// /// - Imports a (none Dart) code file.
+// /// - Attributes:
+// ///   - path= (required) A [DartFilePath] to be imported as a Dart code example. See also [ImportCodeTag] to import none Dart code.
+// ///   - title= (optional) See [TitleAttribute]
+// class ImportDartCodeTag extends Tag {
+//   ImportDartCodeTag(ParentNode? parent, Map<String, dynamic> attributes)
+//       : super(parent, attributes);
+
+//   @override
+//   Future<List<Node>> createChildren() async {
+//     DartFilePath path = attributes[AttributeName.path];
+//     String? title = attributes[AttributeName.title];
+//     var titleAndOrAnchor = TitleAndOrAnchor(this, title, path.toString());
+//     anchor = titleAndOrAnchor.anchor;
+//     var codePrefix = TextNode(this, "$newLine```dart$newLine");
+//     var code = _readCodeFile(path.toFile());
+//     var codeNode = TextNode(this, code);
+//     var codeSuffix = TextNode(this, "$newLine```$newLine");
+//     return [
+//       titleAndOrAnchor,
+//       codePrefix,
+//       codeNode,
+//       codeSuffix,
+//     ];
+//   }
+// }
+
+// /// Recognizes and creates an [ImportDartCodeTag]
+// class ImportDartCodeTagRule extends TagRule {
+//   ImportDartCodeTagRule()
+//       : super('ImportDartCode', [
+//           DartFilePathAttributeRule(),
+//           TitleAttributeRule(),
+//         ]);
+
+//   @override
+//   Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
+//       ImportDartCodeTag(parent, attributes);
+// }
+
+// /// - **{ImportDartDoc path='lib\my_lib.dart|MyClass' title='## My Class'&rcub;**
+// /// - Imports Dart documentation comments from a library member of a dart file.
+// /// - Attributes:
+// ///   - path= (required) A [DartCodePath] to be imported Dart comments.
+// ///   - title= (optional) See [TitleAttribute]
+// class ImportDartDocTag extends Tag {
+//   ImportDartDocTag(ParentNode? parent, Map<String, dynamic> attributes)
+//       : super(parent, attributes);
+
+//   @override
+//   Future<List<Node>> createChildren() async {
+//     DartCodePath path = attributes[AttributeName.path];
+//     String? title = attributes[AttributeName.title];
+//     var titleAndOrAnchor = TitleAndOrAnchor(this, title, path.toString());
+//     anchor = titleAndOrAnchor.anchor;
+//     var documentation = await _readDocumentationComments(parent!, path);
+//     // var documentation = _removeLeadingTripleSlashes(documentationComments);
+//     var codeNode = TextNode(this, documentation);
+//     return [
+//       titleAndOrAnchor,
+//       codeNode,
+//     ];
+//   }
+
+//   Future<String> _readDocumentationComments(
+//       ParentNode parent, DartCodePath path) async {
+//     validate(path);
+
+//     analyzer.LibraryElement library =
+//         await parseLibrary(parent, path.dartFilePath);
+
+//     analyzer.Element foundElement = findAnalyzerElement(library, path);
+
+//     var documentationComments = foundElement.documentationComment ?? '';
+//     String documentation = _removeLeadingTripleSlashes(documentationComments);
+//     validateIfNotEmpty(documentation, path);
+
+//     return documentation;
+//   }
+
+//   static final leadingTripleSlashesExpression = FluentRegex()
+//       .startOfLine()
+//       .whiteSpace(Quantity.zeroOrMoreTimes())
+//       .literal('///')
+//       .whiteSpace(Quantity.atMost(1));
+//   static final lineBreakExpression = FluentRegex().lineBreak();
+
+//   _removeLeadingTripleSlashes(String documentationComments) {
+//     var lines = documentationComments.split(lineBreakExpression);
+//     String doc = '';
+//     lines.forEach((line) {
+//       doc += leadingTripleSlashesExpression.removeFirst(line) + '$newLine';
+//     });
+//     return doc;
+//   }
+
+//   void validate(DartCodePath path) {
+//     validateIfDartFilePathExists(path);
+//     validateIfMemberPathExists(path);
+//   }
+
+//   void validateIfDartFilePathExists(DartCodePath path) {
+//     if (!path.dartFilePath.toFile().existsSync()) {
+//       throw ParserWarning(
+//           'The Dart file could not be found in path attribute value: $path');
+//     }
+//   }
+
+//   void validateIfMemberPathExists(DartCodePath path) {
+//     if (path.dartMemberPath == null) {
+//       throw ParserWarning('No DartMemberPath in path attribute value: $path');
+//     }
+//   }
+
+//   analyzer.Element findAnalyzerElement(
+//       analyzer.Element element, DartCodePath path) {
+//     var visitor = ElementFinder(path.dartMemberPath!);
+//     element.visitChildren(visitor);
+
+//     analyzer.Element? foundElement = visitor.foundElement;
+//     if (foundElement == null) {
+//       throw ParserWarning(
+//           'Dart member: ${path.dartMemberPath} not found in: ${path.dartFilePath}');
+//     }
+//     return foundElement;
+//   }
+
+//   void validateIfMemberFound(
+//       analyzer.Element? foundElement, DartCodePath path) {
+//     if (foundElement == null) {}
+//   }
+
+//   void validateIfNotEmpty(String docComments, DartCodePath path) {
+//     if (docComments.trim().isEmpty) {
+//       throw ParserWarning(
+//           'Dart member: ${path.dartMemberPath} has no Dart documentation comments in: ${path.dartFilePath}');
+//     }
+//   }
+// }
+
+// Future<analyzer.LibraryElement> parseLibrary(
+//     ParentNode parent, ProjectFilePath dartFile) async {
+//   try {
+//     var documentationModel = parent.findParent<DocumentationModel>();
+//     var resolver = documentationModel!.buildStep!.resolver;
+//     var assetId = AssetId(LocalProject.name, dartFile.path);
+//     var library = await resolver.libraryFor(assetId);
+//     return library;
+//   } on Exception catch (e) {
+//     throw ParserWarning('Could not parse library file: ${dartFile.path}',
+//         ParserWarning(e.toString()));
+//   }
+// }
+
+// /// Recognizes and creates an [ImportDartDocTag]
+// class ImportDartDocTagRule extends TagRule {
+//   ImportDartDocTagRule()
+//       : super('ImportDartDoc', [
+//           DartCodePathAttributeRule(),
+//           TitleAttributeRule(),
+//         ]);
+
+//   @override
+//   Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
+//       ImportDartDocTag(parent, attributes);
+// }
+
+// /// Replaces titles (text starting with # or ## or ### etc) with TitleNodes
+// /// This must be done before [TableOfContentsTagRule] so that titles in template files
+// /// can also be found by the [TableOfContentsTag]
+// class TitleRule extends TextParserRule {
+//   TitleRule() : super(createExpression());
+
+//   @override
+//   Future<Node> createReplacementNode(ParentNode parent, RegExpMatch match) =>
+//       Future.value(Title(parent, match.result));
+
+//   static FluentRegex createExpression() => FluentRegex()
+//       .startOfLine()
+//       .whiteSpace(Quantity.zeroOrMoreTimes())
+//       .literal('#', Quantity.between(1, 6))
+//       .whiteSpace(Quantity.zeroOrMoreTimes())
+//       .characterSet(CharacterSet.exclude().addLiterals('#'),
+//           Quantity.oneOrMoreTimes().reluctant)
+//       .endOfLine();
+// }
+
+// /// - **{TableOfContents title='## Table of contents example'&rcub;**
+// /// - Creates a table of contents.
+// /// - Attributes:
+// ///   - title= (optional) See [TitleAttribute]
+// class TableOfContentsTag extends Tag {
+//   TableOfContentsTag(ParentNode? parent, Map<String, dynamic> attributes)
+//       : super(parent, attributes);
+
+//   @override
+//   Future<List<Node>> createChildren() {
+//     String toc = "";
+//     DocumentationModel? model = parent!.findParent<DocumentationModel>();
+//     if (model != null) {
+//       var templates = model.findOrderedMarkdownTemplates();
+//       for (Template template in templates) {
+//         if (template is WikiTemplate && !isWikiHomePage(template)) {
+//           Link titleLink = Link(
+//               parent: parent,
+//               title: template.title,
+//               uri: template.destinationWebUri!);
+//           toc += "$newLine**$titleLink**$newLine$newLine";
+//           List<Title> subTitles = template.findChildren<Title>();
+//           for (Title subTitle in subTitles) {
+//             var title = subTitle.titleWithoutHashes;
+//             var uri = subTitle.anchor.uriToAnchor;
+//             var level = subTitle.level; //1,2,3
+//             if (uri != null && level <= 3) {
+//               Link subTitleLink = Link(parent: parent, title: title, uri: uri);
+//               toc += "${'  ' * (level - 1)}- $subTitleLink$newLine";
+//             }
+//           }
+//         }
+//       }
+//     }
+//     String? titleText = attributes[AttributeName.title];
+//     var tableOfContents = TextNode(this, toc);
+//     if (titleText != null && titleText.trim().isNotEmpty) {
+//       var title = Title(this, titleText);
+//       anchor = title.anchor;
+//       return Future.value([title, tableOfContents]);
+//     } else {
+//       return Future.value([tableOfContents]);
+//     }
+//   }
+
+//   bool isWikiHomePage(Template template) =>
+//       template.sourceFilePath.path.toLowerCase().contains('home.mdt');
+// }
+
+// /// Recognizes and creates an [TableOfContentsTag]
+// class TableOfContentsTagRule extends TagRule {
+//   TableOfContentsTagRule()
+//       : super('TableOfContents', [
+//           TitleAttributeRule(),
+//         ]);
+
+//   @override
+//   Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) =>
+//       TableOfContentsTag(parent, attributes);
+// }
+
+// class TitleAndOrAnchor extends ParentNode {
+//   late final Anchor anchor;
+
+//   TitleAndOrAnchor(ParentNode? parent, String? title, String path)
+//       : super(parent) {
+//     if (title != null) {
+//       var titleNode = Title(this, title);
+//       anchor = titleNode.anchor;
+//       children.add(titleNode);
+//     } else {
+//       anchor = Anchor(this, path);
+//       children.add(anchor);
+//     }
+//   }
+// }
+
+// /// Recognizes and creates a [MitLicenseTag]
+// class MitLicenseTagRule extends TagRule {
+//   MitLicenseTagRule()
+//       : super('mitLicense',
+//             [StringAttributeRule(AttributeName.name, required: true)]);
+
+//   @override
+//   Tag createTagNode(ParentNode parent, Map<String, dynamic> attributes) {
+//     return MitLicenseTag(parent, attributes);
+//   }
+// }
+
+// /// - **{MitLicense name='CopyRight holder'&rcub;**
+// /// - Writes the MIT License with current year and copyright holder
+// /// - Attributes:
+// ///   - name= (required) Copyright holder name.
+// class MitLicenseTag extends Tag {
+//   MitLicenseTag(ParentNode? parent, Map<String, dynamic> attributes)
+//       : super(parent, attributes);
+
+//   String mitLicenseText(String name) => 'MIT License:$newLine$newLine'
+//       'Copyright (c) ${DateTime.now().year} $name$newLine$newLine'
+//       'Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:$newLine$newLine'
+//       'The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.$newLine$newLine'
+//       'THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.';
+
+//   @override
+//   Future<List<Node>> createChildren() {
+//     String name = attributes[AttributeName.name];
+//     return Future.value([TextNode(parent, mitLicenseText(name))]);
+//   }
+// }
+
+// /// Represents a markdown title with a custom Anchor id that can be navigated to using [anchor.uriToAnchor]
+// ///
+// /// Output example:
+// /// <a id='paragraph-titile'></a>
+// /// ## Paragraph Title
+// class Title extends ParentNode {
+//   late final Anchor anchor;
+//   final String title;
+//   final String titleWithoutHashes;
+
+//   /// The title can have leading hashtags to indicate the title level, e.g.:
+//   /// # = Chapter (level=1)
+//   /// ## = Paragraph (level=2)
+//   /// ### = Sub paragraph (level=3)
+//   /// etc... up to 6 levels.
+//   final int level;
+
+//   /// The title can have leading hashtags to indicate the title level, e.g.:
+//   /// # = Chapter
+//   /// ## = Paragraph
+//   /// ### = Sub paragraph
+//   /// etc... up to 6 levels.
+//   Title(ParentNode parent, this.title)
+//       : titleWithoutHashes = createTitleWithoutHashes(title),
+//         level = createLevel(title),
+//         super(parent) {
+//     anchor = Anchor(this, title);
+//     children.add(anchor);
+//     children.add(TitleText(this, title));
+//   }
+
+//   static final FluentRegex _leadingHashExpression = FluentRegex()
+//       .startOfLine()
+//       .group(FluentRegex().whiteSpace(Quantity.zeroOrMoreTimes()).literal('#'),
+//           quantity: Quantity.atMost(6))
+//       .whiteSpace(Quantity.zeroOrMoreTimes());
+
+//   static createLevel(String title) {
+//     String? leadingHashes = _leadingHashExpression.findFirst(title);
+//     if (leadingHashes == null) {
+//       return 7;
+//     } else {
+//       return FluentRegex().literal('#').allMatches(leadingHashes).length;
+//     }
+//   }
+
+//   static createTitleWithoutHashes(String title) {
+//     return _leadingHashExpression.removeFirst(title);
+//   }
+// }
+
+// /// [TitleText] is needed because we can not use [TextNode] inside a [Title].
+// /// This would create a Stackoverflow by the [TitleRule] that keeps replacing
+// /// the [TextNode] inside a [Title]
+// class TitleText extends Node {
+//   final String title;
+
+//   TitleText(ParentNode? parent, this.title) : super(parent);
+
+//   /// Titles must be on a single line to be recognized as a title by the markdown engine
+//   @override
+//   String toString() => '$newLine$title$newLine';
+// }
+
+// /// Represents a HTML anchor point to which you can refer to with a [uriToAnchor]
+// /// https://stackoverflow.com/questions/5319754/cross-reference-named-anchor-in-markdown
+// class Anchor extends Node {
+//   late final String html;
+//   late final String name;
+//   late final Uri? uriToAnchor;
+//   static final firstHyphen = FluentRegex().startOfLine().literal('-');
+//   static final multipleHyphen = FluentRegex().literal('-', Quantity.atLeast(2));
+//   static final whiteSpace = FluentRegex().whiteSpace();
+//   static final otherThanLettersNumbersAndHyphens = FluentRegex().characterSet(
+//       CharacterSet.exclude().addLetters().addDigits().addLiterals('-'));
+
+//   Anchor(ParentNode? parent, String textToChangeToName) : super(parent) {
+//     name = createName(textToChangeToName);
+//     html = createHtml(name);
+//     uriToAnchor = createUriToAnchor(name);
+//   }
+
+//   static String createHtml(String name) => "<a id='$name'></a>";
+
+//   /// Converts a text to a name that can be referred to in a uri.
+//   static String createName(String textToChangeToName) => textToChangeToName
+//       .trim()
+//       .replaceAll(whiteSpace, '-')
+//       .replaceAll(otherThanLettersNumbersAndHyphens, '-')
+//       .replaceAll(multipleHyphen, '-')
+//       .replaceAll(firstHyphen, '')
+//       .toLowerCase();
+
+//   Uri? createUriToAnchor(String name) {
+//     if (parent == null) return null;
+//     Template? markdownTemplate = parent!.findParent<Template>();
+//     if (markdownTemplate == null) return null;
+//     Uri? uri = markdownTemplate.destinationWebUri;
+//     if (uri == null) return null;
+//     uri = uri.replace(fragment: name);
+//     return uri;
+//   }
+
+//   @override
+//   String toString() => html;
+// }
+
+// String _readFile(File file) {
+//   try {
+//     return file.readAsStringSync();
+//   } on Exception catch (e) {
+//     throw ParserWarning(
+//         'Could not read file: ${file.path}.', ParserWarning(e.toString()));
+//   }
+// }
+
+// final _leadingWhiteSpace = FluentRegex()
+//     .startOfLine()
+//     .whiteSpace(Quantity.oneOrMoreTimes())
+//     .multiline(false);
+// final _trailingWhiteSpace =
+//     FluentRegex().whiteSpace(Quantity.oneOrMoreTimes()).endOfLine();
+
+// String _trimWhiteSpace(String text) {
+//   text = _leadingWhiteSpace.removeFirst(text);
+//   var matches = _trailingWhiteSpace.allMatches(text);
+//   if (matches.isNotEmpty) {
+//     var lastMatch = matches.last;
+//     var endPos = lastMatch.end;
+//     if (endPos == text.length) {
+//       text = text.substring(0, lastMatch.start);
+//     }
+//   }
+//   return text;
+// }
+
+// String _readCodeFile(File file) => _trimWhiteSpace(_readFile(file));
